Rename ServiceTest to UnitTest for the PMD checking to pass. Adjust indentation and remove $ from the variables
